Use a timerfd to trigger sending packets
 - Switch from forking off a process and waiting for a timer to expire
   to using a transmitter fork that multiplexes on the timers of all
   packets.
 - Add the GENERIC_LIST helper macro.
 - Add tests for the GENERIC_LIST helper macro.

+++ tests/test_buffer.cpp
@@ -18,7 +18,7 @@ TEST(buffer, zero_size_init) {
   for (i = 0; i < buf->len; ++i) {
     EXPECT_EQ(buf->data[i], 0x00);
   }
-  EXPECT_EQ(i, 9);
+  EXPECT_EQ(i, (size_t)9);
 
   sylkie_buffer_free(buf);
 }
